Adding the fog-bar (fogthecatman)
