+++ mix.exs
@@ -34,8 +34,8 @@ defmodule TwitchDiscovery.Mixfile do
      {:postgrex, ">= 0.0.0"},
      {:phoenix_html, "~> 1.1"},
      {:exredis, ">= 0.1.1"},
-     {:rest_twitch, github: "rockerboo/rest_twitch"},
-     # {:rest_twitch, path: "/home/rockerboo/projects/rest_twitch"},
+     # {:rest_twitch, github: "rockerboo/rest_twitch"},
+     {:rest_twitch, path: "/home/rockerboo/projects/rest_twitch"},
      {:httpoison, "~> 0.7"},
      {:timex, "~> 0.15.0"},
      {:phoenix_live_reload, "~> 0.4", only: :dev},
