+++ web/templates/layout/app.html.eex
@@ -8,6 +8,7 @@
     <!-- load MUI -->
     <link href="//cdn.muicss.com/mui-0.1.14/css/mui.min.css" rel="stylesheet" type="text/css" />
     <script src="//cdn.muicss.com/mui-0.1.14/js/mui.min.js"></script>
+    <script src="//code.jquery.com/jquery-1.11.3.min.js"></script>
 
     <style type="text/css">
     body {
@@ -16,15 +17,22 @@
 
     .fa-bars {
       font-size: 32px;
-      margin: 18px;
+      margin: 7px;
     }
 
-    .video_meta {
+    .app-bar {
+      background: url("http://static-cdn.jtvnw.net/previews-ttv/live_user_lirik-320x180.jpg");
+      background-size: 33%;
+      padding: 10px;
+      padding-left: 16px;
+    }
+
+    #preview_video_meta {
       float: left;
       display: block;
     }
 
-    .video_meta .views, .video_meta .length  {
+    #preview_video_meta .views, #preview_video_meta .length  {
       font-size: 18px;
       padding-right: 10px;
     }
@@ -33,13 +41,252 @@
       margin-bottom: 10px;
     }
 
+    .stream-block img {
+      width: 320px;
+    }
+
+    .inner {
+      margin-bottom: 20px;
+    }
+
+    @media all and (max-width: 400px) {
+      .stream-block img {
+        width: calc(100vw - 75px);
+      }
+    }
+
+#preview {
+  /*height: calc(21vw/21vh*150);*/
+  height: 30vh;
+  min-height: 250px;
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  background: #000;
+  color: #fff;
+}
+
+.preview_off {
+ display: none;
+}
+
+#preview_video {
+  float: left;
+  width: 33.33%;
+  min-width: 33.33%;
+  height: 100%;
+}
+#preview_chat {
+  float: left;
+  width: 33.33%;
+  height: 100%;
+  display: block;
+}
+
+#preview_chat iframe, #preview_video iframe, #preview_chat img, #preview_video img  {
+  width: 100%;
+}
+
+#preview_chat iframe, #preview_chat img, #preview_video iframe {
+  height: 100%;
+}
+
+#preview_meta {
+  float: left;
+  /*width: 33.33vw;*/
+  height: 100%;
+  width: 0;
+  padding: 10px;
+}
+
+#preview_video iframe {
+  width: 100%;
+  height: 100%;
+}
+    @media all and (max-width: 550px) {
+      #preview {
+        height: calc(500/10vh*1.6)px;
+      }
+
+      #preview_video {
+        width: 100vw;
+      }
+
+      #preview_video iframe {
+        width: 100%;
+      }
+
+      #preview_meta {
+        display: none;
+      }
+      #preview_chat {
+        display: none;
+      }
+    }
+
+    /*50/50/0*/
+    @media all and (min-width: 550px) and (max-width: 650px) {
+      #preview_video {
+        width: 50%;
+        display: block;
+      }
+
+      #preview_chat {
+        width: 50%;
+        display: block;
+      }
+
+      #preview_meta {
+        width: 0;
+        display: none;
+      }
+    }
+
+    /*33/66/0*/
+    @media all and (min-width: 650px) and (max-width: 830px) {
+      #preview_video {
+        width: 33.33%;
+        display: block;
+      }
+
+      #preview_chat {
+        width: 66.66%;
+        display: block;
+      }
+
+      #preview_meta {
+        width: 0;
+        display: none;
+      }
+    }
+
+    /*33/33/33*/
+    @media all and (min-width: 830px) and (max-width: 950px) {
+      #preview_video {
+        width: 33.33%;
+        display: block;
+      }
+
+      #preview_chat {
+        width: 33.33%;
+        display: block;
+      }
+
+      #preview_meta {
+        width: 33.33%;
+        display: block;
+      }
+    }
+
+    /*33/33/33*/
+    @media all and (min-width: 950px) and (max-width: 1050px) {
+      #preview_video {
+        width: 33.33%;
+        display: block;
+      }
+
+      #preview_chat {
+        width: 33.33%;
+        display: block;
+      }
+
+      #preview_meta {
+        width: 33%;
+        display: block;
+      }
+    }
     </style>
+    <script type="text/javascript">
+
+
+function playerContainer(channel) {
+  var container = document.createElement('iframe')
+  container.setAttribute("id", "twitch_player")
+  container.setAttribute("type", "text/html")
+  container.setAttribute("src", "http://www.twitch.tv/" + channel + "/embed")
+  container.setAttribute("frameborder", "0")
+
+  return container;
+}
+
+function chatContainer(channel) {
+  var container = document.createElement('iframe')
+  container.setAttribute("id", "chat_embed")
+  container.setAttribute("type", "text/html")
+  container.setAttribute("src", "http://www.twitch.tv/" + channel + "/chat")
+  container.setAttribute("frameborder", "0")
+  container.setAttribute("scrolling", "no")
+
+  return container;
+}
+
+function twitchPlayer(channel) {
+    container = playerContainer(channel)
+
+    var player = document.getElementById("preview_video")
+
+    player.appendChild(container);
+}
+
+function twitchChat(channel) {
+    container = chatContainer(channel)
+
+    var chat = document.getElementById("preview_chat")
+
+    chat.appendChild(container);
+}
+
+function clearPreview() {
+  document.getElementById("preview_chat").innerHTML = ""
+  video = document.getElementById("preview_video")
+
+  video.innerHTML = ""
+  video.parentNode.classList.add("preview_off")
+}
+
+window.onload = function () {
+  $("a.preview").click(function(event) {
+      event.preventDefault()
+
+      channel = event.currentTarget.getAttribute("data-channel")
+
+      console.log(event)
+      console.log(channel)
+
+      clearPreview()
+
+      twitchPlayer(channel)
+      twitchChat(channel)
+
+      document.getElementById("preview_video").parentNode.classList.remove("preview_off")
+
+      // .getAttribute("data-course-id"))
+
+      // twitchPlayer()
+      // $(this).hide("slow");
+  });
+}
+    </script>
   </head>
   <body>
-    <div class="mui-appbar">
-       <a href="/"><i class="fa fa-bars"></i></a>
+    <div class="app-bar mui-appbar">
+        <!-- <i class="fa fa-bars"></i> -->
+       <a href="/"><img src="http://elirc.rockerboo.net/assets/images/discovery.svg" style="width: 225px" /></a>
     </div>
+    <div class="inner">
       <%= @inner %>
+    </div>
+    <div id="preview" class="preview_off">
+      <div></div>
+      <div style="clear:both;"><span></span></div>
+      <div id="preview_video">
+      </div>
+      <div id="preview_chat">
+      </div>
+      <div id="preview_meta">
+        Meta Information, follow links, schedule
+      </div>
     </div>
   </body>
 </html>
